+++ src/frontend/ast.ml
@@ -374,7 +374,7 @@ and function_pointer_declarator = {
   signature_parameters : function_parameter list;
   signature_empty_parameter_entries : empty_parameter_entry list;
   signature_variadic : variadic_marker option;
-  signature_closing_parenthesis : location;
+  signature_closing_parenthesis : location option;
   function_pointer_location : location;
 }
 
@@ -490,7 +490,7 @@ type function_prototype = {
   parameters : function_parameter list;
   empty_parameter_entries : empty_parameter_entry list;
   variadic : variadic_marker option;
-  closing_parenthesis : location;
+  closing_parenthesis : location option;
   semicolon : location option;
   location : location;
 }
@@ -869,7 +869,7 @@ type function_definition = {
   parameters : function_parameter list;
   empty_parameter_entries : empty_parameter_entry list;
   variadic : variadic_marker option;
-  closing_parenthesis : location;
+  closing_parenthesis : location option;
   body : statement option;
   location : location;
 }