+++ test/test_stateful_exe_cli.ml
@@ -120,6 +120,13 @@ let function_versions_source =
  StreamPrint("%d;",SavedBefore()+SavedInner()+F()-108);
 }|}
 
+let variadic_termination_source =
+  {|#exe {
+ I64 F(I64 n=40,... { I64 v=n+argc+argv[0]; return v; }
+ #exe { I64 Saved() { return F(,1); } }
+ StreamPrint("%d;",Saved());
+}|}
+
 let parameter_delimiters_source =
   "I64 F(;;I64 n=40,;;I64 m=2,;;){return n+m;};F();"
 
@@ -252,6 +259,8 @@ let () =
       (function_versions_source, "aot", 73, 6);
       (parameter_delimiters_source, "jit", 22, 6);
       (parameter_delimiters_source, "aot", 20, 6);
+      (variadic_termination_source, "jit", 58, 3);
+      (variadic_termination_source, "aot", 60, 3);
     ];
   List.iter
     (fun source ->